requirements.txt
numpy
scipy
scikit-image
Pillow
tifffile
matplotlib
tqdm
tkinterdnd2
numba